Add shortcut buttons for controlling model view
Add buttons for aligning and rotating model, as well as selecting all atoms.
The user can specify the angle to be rotated with spin button. The rotation button images still need some improvement but hopefully are intelligible enough.
Also changed the grouping of buttons slightly to be more "intuitive" (?).
The new alignment buttons render the Model:Viewing buttons redundant, but the latter can be kept for the time being (although the menu buttons are easier to use).
Also added shortcut key combination for "unhiding" all atoms, as well as notice of number of hidden atoms in model viewer. 
Also fixed several spelling mistakes in comments.