Allow editing of element, name and SOF of atom
Most software packages and structure formats distinguish between atom names and element symbols. Allowing these to be modified separately in the Model:Editing window provides greater control to the user for when loading and building models.
Another issue related to compatibility is that, in Gulp and Marvin, regions are labelled from 1 to 4, whereas Gdis uses 0 to 3. To avoid any possible confusion for the user, Model: Editing now displays and accepts inputs 1 to 4 (although internally 0 to 3 are used, e.g., when indexing arrays, and so on).